+++ driver/src/test/java/oracle/nosql/driver/QueryTest.java
@@ -127,8 +127,9 @@ public class QueryTest extends ProxyTestBase {
 
     @Override
     public void afterTest() throws Exception {
-        handle.queryIterable(new QueryRequest().setStatement("DELETE FROM " +
-            "testTable"));
+        QueryRequest qreq = new QueryRequest();
+        qreq.setStatement("DELETE FROM testTable");
+        handle.queryIterable(qreq);
         tableOperation(handle, "DROP TABLE testTable", null);
         super.afterTest();
 
@@ -164,16 +165,16 @@ public class QueryTest extends ProxyTestBase {
         /*
          * Perform an update query
          */
-        try (QueryRequest queryRequest =
-            new QueryRequest().setStatement(updateQuery)) {
+        try (QueryRequest queryRequest = new QueryRequest()) {
+            queryRequest.setStatement(updateQuery);
             handle.query(queryRequest);
         }
 
         /*
          * Use a simple get query to validate the update
          */
-        try (QueryRequest queryRequest = new QueryRequest()
-            .setStatement(getQuery)) {
+        try (QueryRequest queryRequest = new QueryRequest()) {
+            queryRequest.setStatement(getQuery);
             QueryResult queryRes = handle.query(queryRequest);
             assertEquals(1, queryRes.getResults().size());
             assertEquals("joe",
@@ -1018,13 +1019,15 @@ public class QueryTest extends ProxyTestBase {
             "(1, 15, \"myname\", 23, \"WI\", 2500, [], \"" +
             longString + "\")";
 
-        req = new QueryRequest().setStatement(query);
+        req = new QueryRequest();
+        req.setStatement(query);
         ret = handle.query(req);
 
         assertTrue(ret.getResults().size() == 1);
 
         query = "select sid, id, name from testTable where id = 15";
-        req = new QueryRequest().setStatement(query);
+        req = new QueryRequest();
+        req.setStatement(query);
         ret = handle.query(req);
         assertTrue(ret.getResults().size() == 1);
         MapValue res = ret.getResults().get(0);
@@ -1056,8 +1059,8 @@ public class QueryTest extends ProxyTestBase {
             .setVariable("$sid", new IntegerValue(0))
             .setVariable("$id", new IntegerValue(1));
 
-        try(QueryRequest req =
-                new QueryRequest().setPreparedStatement(prepRet)) {
+        try(QueryRequest req = new QueryRequest()) {
+            req.setPreparedStatement(prepRet);
             QueryResult res = handle.query(req);
             assertNotNull(res.getResults());
         }
@@ -1107,8 +1110,8 @@ public class QueryTest extends ProxyTestBase {
                 .setVariable("$sid", new IntegerValue(0))
                 .setVariable("$id", new IntegerValue(1));
 
-            try (QueryRequest req = new QueryRequest()
-                                       .setPreparedStatement(prepRet)) {
+            try (QueryRequest req = new QueryRequest()) {
+                req.setPreparedStatement(prepRet);
                 total++;
                 verbose("Running query #" + total + "...");
                 QueryResult res = handle.query(req);
@@ -1181,14 +1184,16 @@ public class QueryTest extends ProxyTestBase {
             fail("query should have failed");
         } catch (IllegalArgumentException iae) {}
 
-        queryReq = new QueryRequest().setStatement("random string");
+        queryReq = new QueryRequest();
+        queryReq.setStatement("random string");
         try {
             handle.query(queryReq);
             fail("query should have failed");
         } catch (IllegalArgumentException iae) {}
 
         /* Try a query that requires external variables that are missing */
-        queryReq = new QueryRequest().setStatement(queryWithVariables);
+        queryReq = new QueryRequest();
+        queryReq.setStatement(queryWithVariables);
         try {
             handle.query(queryReq);
             fail("query should have failed");
@@ -1197,7 +1202,8 @@ public class QueryTest extends ProxyTestBase {
 
         prepReq = new PrepareRequest().setStatement(queryWithVariables);
         PrepareResult prepRes = handle.prepare(prepReq);
-        queryReq = new QueryRequest().setPreparedStatement(prepRes);
+        queryReq = new QueryRequest();
+        queryReq.setPreparedStatement(prepRes);
         try {
             handle.query(queryReq);
             fail("query should have failed");
@@ -1210,7 +1216,8 @@ public class QueryTest extends ProxyTestBase {
         PreparedStatement prepStmt = prepRes.getPreparedStatement();
         prepStmt.setVariable("sid", new IntegerValue(9));
         prepStmt.setVariable("id", new IntegerValue(3));
-        queryReq = new QueryRequest().setPreparedStatement(prepRes);
+        queryReq = new QueryRequest();
+        queryReq.setPreparedStatement(prepRes);
         try {
             handle.query(queryReq);
             fail("query should have failed");
@@ -1223,7 +1230,8 @@ public class QueryTest extends ProxyTestBase {
         prepStmt = prepRes.getPreparedStatement();
         prepStmt.setVariable("$sid", new DoubleValue(9.1d));
         prepStmt.setVariable("$id", new IntegerValue(3));
-        queryReq = new QueryRequest().setPreparedStatement(prepRes);
+        queryReq = new QueryRequest();
+        queryReq.setPreparedStatement(prepRes);
         try {
             handle.query(queryReq);
             fail("query should have failed");
@@ -1239,7 +1247,8 @@ public class QueryTest extends ProxyTestBase {
         } catch (TableNotFoundException tnfe) {
         }
 
-        queryReq = new QueryRequest().setStatement(query);
+        queryReq = new QueryRequest();
+        queryReq.setStatement(query);
         try {
             handle.query(queryReq);
             fail("query should have failed");
@@ -1255,7 +1264,8 @@ public class QueryTest extends ProxyTestBase {
         } catch (IllegalArgumentException iae) {
         }
 
-        queryReq = new QueryRequest().setStatement(query);
+        queryReq = new QueryRequest();
+        queryReq.setStatement(query);
         try {
             handle.query(queryReq);
             fail("query should have failed");
@@ -1271,14 +1281,16 @@ public class QueryTest extends ProxyTestBase {
         } catch (IllegalArgumentException iae) {
         }
 
-        queryReq = new QueryRequest().setStatement(query);
+        queryReq = new QueryRequest();
+        queryReq.setStatement(query);
         try {
             handle.query(queryReq);
             fail("query should have failed");
         } catch (IllegalArgumentException iae) {
         }
 
-        queryReq = new QueryRequest().setStatement(query);
+        queryReq = new QueryRequest();
+        queryReq.setStatement(query);
         try {
             queryReq.setLimit(-1);
             handle.query(queryReq);
@@ -1566,8 +1578,8 @@ public class QueryTest extends ProxyTestBase {
         PrepareResult prepRet = handle.prepare(prepReq);
         assertNotNull(prepRet.getPreparedStatement());
 
-        try(QueryRequest qreq =
-                new QueryRequest().setPreparedStatement(prepRet)) {
+        try(QueryRequest qreq = new QueryRequest()) {
+            qreq.setPreparedStatement(prepRet);
             QueryResult qres = handle.query(qreq);
             assertEquals(10, qres.getResults().size());
 
@@ -1663,8 +1675,9 @@ public class QueryTest extends ProxyTestBase {
         /* Load rows to table */
         loadRowsToScanTable(10, 10, 1);
 
-        try (QueryRequest queryReq = new QueryRequest().
-            setStatement("select * from testTable where id = 1 and sid = 1")) {
+        try (QueryRequest queryReq = new QueryRequest()) {
+            queryReq.setStatement(
+                "select * from testTable where id = 1 and sid = 1");
 
             QueryResult queryRes = handle.query(queryReq);
 
@@ -1722,8 +1735,8 @@ public class QueryTest extends ProxyTestBase {
         /*
          * Ensure that this query completes
          */
-        try (QueryRequest queryReq = new QueryRequest().
-            setStatement("select * from " + name)) {
+        try (QueryRequest queryReq = new QueryRequest()) {
+            queryReq.setStatement("select * from " + name);
             int numRes = 0;
             do {
                 QueryResult queryRes = handle.query(queryReq);
@@ -1787,10 +1800,12 @@ public class QueryTest extends ProxyTestBase {
                 "set t.data = " + data + "where id = 1 returning id";
 
             /* insert, then update */
-            QueryRequest req = new QueryRequest().setStatement(iquery);
+            QueryRequest req = new QueryRequest();
+            req.setStatement(iquery);
             QueryResult res = handle.query(req);
             assertEquals(1, res.getResults().get(0).get("id").getInt());
-            req = new QueryRequest().setStatement(uquery);
+            req = new QueryRequest();
+            req.setStatement(uquery);
             res = handle.query(req);
             assertEquals(1, res.getResults().get(0).get("id").getInt());
         }
@@ -1799,7 +1814,8 @@ public class QueryTest extends ProxyTestBase {
         final String squery = "select * from " + tableName +
             " t where t.data.data = " + genString(15000);
 
-        try (QueryRequest req = new QueryRequest().setStatement(squery)) {
+        try (QueryRequest req = new QueryRequest()) {
+            req.setStatement(squery);
             handle.query(req);
             fail("Query should have failed");
         } catch (IllegalArgumentException iae) {
@@ -1891,8 +1907,8 @@ public class QueryTest extends ProxyTestBase {
         pstmt.setVariable("$id", new IntegerValue(id));
         pstmt.setVariable("$info", arrVal);
 
-        try (QueryRequest req = new QueryRequest().setPreparedStatement(pstmt))
-        {
+        try (QueryRequest req = new QueryRequest()) {
+            req.setPreparedStatement(pstmt);
             handle.query(req);
             fail("Expected IAE");
         } catch (IllegalArgumentException ex) {
@@ -1907,8 +1923,8 @@ public class QueryTest extends ProxyTestBase {
         pstmt.setVariable("$id", new IntegerValue(id));
         pstmt.setVariable("$info", arrVal);
 
-        try (QueryRequest req = new QueryRequest().setPreparedStatement(pstmt))
-        {
+        try (QueryRequest req = new QueryRequest()) {
+            req.setPreparedStatement(pstmt);
             handle.query(req);
             fail("Expected IAE");
         } catch(IllegalArgumentException ex) {
@@ -1920,9 +1936,8 @@ public class QueryTest extends ProxyTestBase {
         /* Load rows to table */
         loadRowsToScanTable(3, 2, 1);
 
-        QueryRequest qreq = new QueryRequest()
-            .setStatement("select * from testTable")
-            .setLimit(3);
+        QueryRequest qreq = new QueryRequest();
+        qreq.setStatement("select * from testTable").setLimit(3);
 
         int count = 0;
         try (QueryIterableResult qir = handle.queryIterable(qreq)) {
@@ -1965,9 +1980,8 @@ public class QueryTest extends ProxyTestBase {
             totalReadUnits = 0, totalWriteUnits = 0;
         RetryStats totalRetryStats = null;
 
-        try (QueryRequest qreq =
-            new QueryRequest().setStatement(query).setLimit(3)) {
-
+        try (QueryRequest qreq = new QueryRequest()) {
+            qreq.setStatement(query).setLimit(3);
 
             QueryResult qres;
 
@@ -1993,8 +2007,9 @@ public class QueryTest extends ProxyTestBase {
             while (!qreq.isDone());
         }
 
-        QueryIterableResult qires = handle.queryIterable(
-            new QueryRequest().setStatement(query).setLimit(3));
+        QueryRequest qreq = new QueryRequest();
+        qreq.setStatement(query).setLimit(3);
+        QueryIterableResult qires = handle.queryIterable(qreq);
         Set<MapValue> actualSet = new HashSet<>();
 
         for (MapValue qiRow : qires) {
@@ -2035,9 +2050,11 @@ public class QueryTest extends ProxyTestBase {
      * the same as regular query() results.
      */
     private void checkQueryIterableOrdered(String query) {
-        QueryRequest qireq = new QueryRequest().setStatement(query);
+        QueryRequest qireq = new QueryRequest();
+        qireq.setStatement(query);
         try (QueryIterableResult qires = handle.queryIterable(qireq);
-             QueryRequest qreq = new QueryRequest().setStatement(query)) {
+             QueryRequest qreq = new QueryRequest()) {
+            qreq.setStatement(query);
 
             Iterator<MapValue> qiIter = qires.iterator();
 
@@ -2105,9 +2122,8 @@ public class QueryTest extends ProxyTestBase {
         loadRowsToScanTable(numMajor, numPerMajor, 1);
 
         /* Get all results and try to use the iterator again */
-        QueryRequest qr = new QueryRequest()
-            .setLimit(4)
-            .setStatement("select * from testTable");
+        QueryRequest qr = new QueryRequest();
+        qr.setLimit(4).setStatement("select * from testTable");
 
         int count = 0;
         QueryIterableResult qires = handle.queryIterable(qr);
@@ -2162,15 +2178,16 @@ public class QueryTest extends ProxyTestBase {
         pstmt.setVariable("$id", new IntegerValue(id));
         pstmt.setVariable("$info", info);
 
-        try (QueryRequest req =
-                 new QueryRequest().setPreparedStatement(pstmt)) {
+        try (QueryRequest req = new QueryRequest()) {
+            req.setPreparedStatement(pstmt);
             QueryResult ret = handle.query(req);
             assertEquals(1, ret.getResults().get(0).asMap()
                 .get("NumRowsInserted").getInt());
         }
 
         String stmt = "select info from " + tableName + " where id = " + id;
-        try (QueryRequest req = new QueryRequest().setStatement(stmt)) {
+        try (QueryRequest req = new QueryRequest()) {
+            req.setStatement(stmt);
             QueryResult ret = handle.query(req);
             assertEquals(1, ret.getResults().size());
             assertEquals(expInfo, ret.getResults().get(0).get("info"));
@@ -2207,11 +2224,10 @@ public class QueryTest extends ProxyTestBase {
                               int recordKB,
                               Consistency consistency) {
 
-        try (final QueryRequest queryReq = new QueryRequest()
-            .setStatement(statement)
-            .setLimit(numLimit)
-            .setConsistency(consistency)
-            .setMaxReadKB(sizeLimit)) {
+        try (final QueryRequest queryReq = new QueryRequest()) {
+            queryReq.setStatement(statement).setLimit(numLimit)
+                .setConsistency(consistency)
+                .setMaxReadKB(sizeLimit);
 
             if (consistency != null) {
                 queryReq.setConsistency(consistency);